+++ CHANGES.md
@@ -6,6 +6,7 @@
 * Cursor origin moved from (1, 1) to (0, 0)
 * Query a terminal for file descriptors it is connected to
 * Export more fine-grained types for input; #key -> #special
+* Add attr and image equality
 
 0.1.1 (2016-02-09):
 * Term.input -> Term.event